\name{distLprint}
\alias{distLprint}
\title{print dlf objects}
\description{print list objects created in this package}
\usage{distLprint(dlf, digits=1)}
\arguments{
  \item{dlf}{List as explained in \code{\link{extremeStat}}}
  \item{digits}{number of digits \code{\link{round}}ed to. DEFAULT: 1}
}
\value{none, prints via \code{\link{message}}.}
\author{Berry Boessenkool, \email{berry-b@gmx.de}, March 2015}
\seealso{\code{\link{extremeStat}}}
\examples{
# see 
?distLextreme
}
\keyword{list}
\keyword{methods}
\keyword{print}
