% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{extremeStat-deprecated}
\alias{extremeStat-deprecated}
\alias{distLextremePlot}
\alias{distLgofPlot}
\alias{distLplot}
\alias{distLprint}
\alias{distLgof}
\title{Deprecated Functions}
\usage{
distLextremePlot(...)

distLgof(...)

distLgofPlot(...)

distLplot(...)

distLprint(...)
}
\arguments{
\item{\dots}{Ignored arguments}
}
\description{
With version 1.3.0 (Jan 2017), a number of functions have been renamed and restructured.
The computing functions don't plot anymore.
This is changed to reduce number of arguments in functions
and increase reproducibility by explicitly creating dlf objects.
Renaming has been done to create a consistent and short naming pattern.
See the new structure in \code{\link{extremeStat}}.
}
\seealso{
\code{\link{extremeStat}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2017
}
