% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeStat-package.R
\docType{package}
\name{extremeStat-package}
\alias{extremeStat-package}
\title{extremeStat: Extreme Value Statistics and Quantile Estimation}
\description{
Fit, plot and compare several (extreme value) distribution functions. Compute (truncated) distribution quantile estimates and plot return periods on a linear scale. On the fitting method, see Asquith (2011): Distributional Analysis with L-moment Statistics [...] ISBN 1463508417.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brry/extremeStat}
  \item Report bugs at \url{https://github.com/brry/extremeStat/issues}
}

}
\author{
\strong{Maintainer}: Berry Boessenkool \email{berry-b@gmx.de}

}
\keyword{internal}
