% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeStat-package.R
\docType{package}
\name{extremeStat}
\alias{extremeStat}
\alias{extremeStat-package}
\title{Extreme value statistics on a linear scale}
\description{
Fit (via L moments), plot (on a linear scale) and compare (by goodness of fit)
several (extreme value) distributions.
Compute high quantiles even in small samples and estimate extrema at given return periods.\cr
Open the \href{https://cran.r-project.org/package=extremeStat/vignettes/extremeStat.html}{Vignette} 
for an introduction to the package. \code{vignette("extremeStat")}\cr
This package heavily relies on and thankfully acknowledges the package \code{lmomco} by WH Asquith.
}
\details{
The main functions in the extremeStat package are:
\tabular{lll}{
\code{\link{distLweights}} -> \tab \code{\link{distLgof}} \tab \code{\link{plotLgof}} \cr
                              \tab \code{\link{distLfit}} \tab \code{\link{plotLfit}} \cr
\code{\link{q_gpd}}, \code{\link{q_weighted}} \tab \code{\link{distLquantile}} \tab \code{link{plotLquantile}} \cr
                              \tab \code{\link{distLextreme}} \tab \code{\link{plotLextreme}} \cr
                              \tab \code{\link{distLexBoot}} \tab \cr
}

The common object to share between functions is a list (\code{dlf}) with:\cr
\tabular{ll}{
\code{dat}       \tab numeric vector with (extreme) values \cr
\code{datname}   \tab character string for main, xlab etc \cr
\code{parameter} \tab list (usually of length 17 if \code{speed=TRUE}) with parameters of each distribution\cr
\code{gof}       \tab dataframe with 'Goodness of Fit' measures, sorted by RMSE of theoretical and empirical cumulated density\cr
\code{returnlev} \tab dataframe with values of distributions for given return periods (\code{RPs}). This element is only added in \code{\link{distLextreme}}\cr
\code{RP___}     \tab Return periods according to plotting positions, see below. \cr
\code{coldist}   \tab Colors for plotting, added in \code{\link{plotLfit}}\cr
\code{truncate}  \tab Truncation percentage, only relevant for \code{\link{distLquantile}}\cr
\code{quant}     \tab Quantile estimation from \code{\link{distLquantile}}\cr
}
It can be printed with \code{\link{printL}}.\cr
PP:\cr
Plotting positions are not used for fitting distributions, but for plotting only\cr
The ranks of ascendingly sorted extreme values are used to compute the probability of non-exceedence Pn:\cr
\code{Pn_w <-  Rank      /(n+1)       # Weibull}\cr
\code{Pn_g <- (Rank-0.44)/(n+0.12)    # Gringorton (taken from lmom:::evplot.default)}\cr
Finally: RP = Returnperiod = recurrence interval = 1/P_exceedence = 1/(1-P_nonexc.), thus:\cr
\code{RPweibull = 1/(1-Pn_w)} and analogous for gringorton.\cr

This package was motivated by my need to compare the fits of several distributions to data.
It was originally triggered by a flood estimation assignment we had in class 2012,
and it bothered me that we just assumed the gumbel distribution would fit the data fine.
}
\examples{

data(annMax) # annual discharge maxima from a stream in Austria
plot(annMax, type="l")
dle <- distLextreme(annMax)
dle$returnlev

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2014-2016
}
\seealso{
If you are looking for more detailed (uncertainty) analysis, eg confidence intervals,
check out the package \code{extRemes}, especially the function \code{\link[extRemes]{fevd}}.
\url{https://cran.r-project.org/package=extRemes}\cr
Intro slides: \url{http://sites.lsa.umich.edu/eva2015/wp-content/uploads/sites/44/2015/06/Intro2EVT.pdf}\cr
Parameter fitting and distribution functions: \url{https://cran.r-project.org/package=lmomco}\cr
Distributions: \url{https://www.rmetrics.org/files/Meielisalp2009/Presentations/Scott.pdf}
and: \url{https://cran.r-project.org/view=Distributions} \cr
R in Hydrology: \url{http://abouthydrology.blogspot.de/2012/08/r-resources-for-hydrologists.html}\cr
}
\keyword{documentation}
\keyword{package}

