% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLexBoot.R
\name{plotLexBoot}
\alias{plotLexBoot}
\title{Bootstrapping uncertainty intervals for return periods}
\usage{
plotLexBoot(dlf, selection = NULL, add = FALSE, log = TRUE, ...)
}
\arguments{
\item{dlf}{\code{dlf} object, as returned by \code{\link{distLexBoot}}}

\item{selection}{Character vector with distribution function names to be used. 
Suggested to keep this low. DEFAULT: NULL}

\item{add}{Add to existing plot? DEFAULT: FALSE}

\item{log}{Plot on a logarithmic axis. DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{plotLextreme}}.
If add=TRUE, they are instead passed to 
\code{berryFunctions::\link{ciBand}}}
}
\value{
invisible dlf object, see \code{\link{printL}}
}
\description{
plot bootstrap uncertainty intervals for \code{\link{plotLextreme}}.
}
\examples{
# see distLexBoot

}
\seealso{
\code{\link{distLexBoot}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2016
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{dplot}
\keyword{hplot}
\keyword{montecarlo}
\keyword{ts}
