% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_extRemes.R
\name{q_extRemes}
\alias{q_extRemes}
\title{GPD quantile of sample}
\usage{
q_extRemes(x, probs, truncate, returnlist = FALSE, undertruncNA = TRUE,
  quiet = FALSE, method = "MLE", ...)
}
\arguments{
\item{x}{Vector with values. NAs are silently ignored}

\item{probs}{Probabilities of truncated (Peak over treshold) quantile}

\item{truncate}{Truncation percentage (proportion of sample discarded)}

\item{returnlist}{Return result from \code{extRemes::\link[extRemes]{fevd}} with the quantiles added to the list as element \code{quant}? DEFAULT: FALSE}

\item{undertruncNA}{Return NAs for probs below truncate? Highly recommended to leave this at the DEFAULT: TRUE}

\item{quiet}{Should messages be suppressed? DEFAULT: FALSE}

\item{method}{method in \code{extRemes::\link[extRemes]{fevd}}, one of c("MLE", "GMLE", "Bayesian", "Lmoments"). Bayesian is slow! DEFAULT: "MLE"}

\item{\dots}{Further arguments passed to \code{extRemes::\link[extRemes]{fevd}}}
}
\value{
if(returnlist): fevd list with element \code{quant} added \cr
else: Vector of quantile estimates for each value of probs
}
\description{
Compute quantile of General Pareto Distribution fitted to sample
by peak over treshold (POT) method declaring treshold from truncation proportion
using the extRemes package (Weather and Climate Applications of Extreme Value Analysis)
}
\examples{

data(annMax)
p <- c(0.5, 0.8, 0.9, 0.99, 0.999)
d <- q_extRemes(annMax, truncate=0,   p=p); d
d <- q_extRemes(annMax, truncate=0,   p=p, returnlist=TRUE); d
d <- q_extRemes(annMax, truncate=0.6, p=p); d
d <- q_extRemes(annMax, truncate=0.6, p=p, under=FALSE); d
d <- q_extRemes(annMax, truncate=0.6, p=p, returnlist=TRUE); d
d <- q_extRemes(annMax, truncate=1,   p=p); d
d <- q_extRemes(annMax, truncate=1,   p=p, returnlist=TRUE); d
d <- q_extRemes(annMax, truncate=1,   p=p, under=FALSE); d

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2016
}
\seealso{
\code{\link{distLquantile}}, \code{extRemes::\link[extRemes]{fevd}}, \code{\link{q_evir}}, \code{\link{q_ismev}}, \code{\link{q_fExtremes}}
}
\keyword{distribution}
\keyword{robust}
\keyword{univar}

