% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_gpd.R
\name{q_gpd}
\alias{q_gpd}
\title{GPD quantile of sample}
\usage{
q_gpd(x, probs = c(0.8, 0.9, 0.99), truncate = 0,
  threshold = berryFunctions::quantileMean(x, truncate),
  package = "extRemes", method = NULL, returnlist = FALSE,
  undertruncNA = TRUE, quiet = FALSE, suppresswarnings = quiet, ...)
}
\arguments{
\item{x}{Vector with numeric values. NAs are silently ignored.}

\item{probs}{Probabilities of truncated (Peak over treshold) quantile. DEFAULT: c(0.8,0.9,0.99)}

\item{truncate}{Truncation percentage (proportion of sample discarded). DEFAULT: 0}

\item{threshold}{POT cutoff value. If you want correct percentiles, set this only via truncate, see Details. DEFAULT: \code{\link[berryFunctions]{quantileMean}(x, truncate)}}

\item{package}{Character string naming package to be used. One of c("evir","extRemes","fExtremes","ismev"). DEFAULT: "extRemes"}

\item{method}{\code{method} passed to the fitting function, if applicable. Defaults are internally specified, depending on \code{package}, if left to the DEFAULT: NULL.}

\item{returnlist}{Return result from the fitting funtion (if provided) with the quantiles added to the list as element \code{quant}. DEFAULT: FALSE}

\item{undertruncNA}{Return NAs for probs below truncate? Highly recommended to leave this at the DEFAULT: TRUE}

\item{quiet}{Should messages from this function be suppressed? DEFAULT: FALSE}

\item{suppresswarnings}{Should warnings be suppressed via \code{\link{options}(warn=-1)}? The usual type of warning is: NAs produced in log(...). DEFAULT: quiet}

\item{\dots}{Further arguments passed to the fitting funtion listed in section Details.}
}
\value{
if(returnlist): list with element \code{quant} added \cr
else: Named vector of quantile estimates for each value of \code{probs}
}
\description{
Compute quantile of General Pareto Distribution fitted to sample by peak over treshold (POT) method
using treshold from truncation proportion,
comparing several R packages doing this
}
\details{
Depending on the value of "package", this fits the GPD using \cr
\code{evir::\link[evir]{gpd}}\cr
\code{extRemes::\link[extRemes]{fevd}}\cr
\code{fExtremes::\link[fExtremes]{gpdFit}}\cr
\code{ismev::\link[ismev]{gpd.fit}}\cr
The \code{method} defaults (and other possibilities) are \cr
evir: "pwm" (probability-weighted moments), or "ml" (maximum likelihood) \cr
extRemes: "MLE", or "GMLE", "Bayesian", "Lmoments" \cr
fExtremes: "pwm", or "mle"\cr
ismev: none, only Maximum-likelihood fitting implemented \cr
The Quantiles are always given with \code{probs} in regard to the full (uncensored) sample.
If e.g. truncate is 0.90, the distribution function is fitted to the top 10\% of the sample.
The 95th percentile of the full sample is equivalent to the 50\% quantile of the subsample acutally used for fitting.
For computation, the probabilities are internally updated with \code{p2=(p-t)/(1-t)}
but labelled with the original \code{p}.
If you truncate 90\% of the sample, you cannot compute the 70th percentile anymore,
thus \code{undertruncNA} should be left to TRUE. \cr
If not exported by the packages, the quantile functions are extracted from their current (Feb 2016) source code.
}
\examples{
 # q_gpd examples are currently in external file!
 stop("Berry, put the q_gpd examples back here!")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2016
}
\seealso{
\code{\link{distLquantile}} which compares results for all packages
}
\keyword{distribution}
\keyword{robust}
\keyword{univar}

