% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_weighted.R
\name{q_weighted}
\alias{q_weighted}
\title{Compute weighted averages of quantile estimates}
\usage{
q_weighted(quant, weights = distLweights(quant, ...), ...)
}
\arguments{
\item{quant}{Data.frame as in \code{\link{distLquantile}} output.}

\item{weights}{Data.frame as in \code{\link{distLweights}} output.}

\item{\dots}{Arguments passed to \code{\link{distLweights}} 
like weightc, onlydn=FALSE. order will be ignored, as 
q_weighted only adds/changes the rows weighted*.}
}
\value{
data.frame with rows "weighted*" added.
}
\description{
Compute weighted averages of quantile estimates
}
\examples{
x <- data.frame(A=1:5, RMSE=runif(5))
distLweights(x)

q_weighted(x)
q_weighted(x,  distLweights(x, weightc=c("1"=3, "3"=5), order=FALSE)  )

x <- rexp(190)
d <- distLquantile(x)
d2 <- q_weighted(d)
stopifnot(all(d==d2, na.rm=TRUE))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2016
}
\seealso{
\code{\link{distLquantile}}
}
\keyword{distribution}

