% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLgof.R
\name{distLgof}
\alias{distLgof}
\title{Quality of distribution fits}
\usage{
distLgof(dlf, order = TRUE, plot = TRUE, progbars = length(dlf$dat) > 200,
  ks = FALSE, weightc = NA, quiet = FALSE, ...)
}
\arguments{
\item{dlf}{List as returned by \code{\link{distLfit}}, containing the elements
\code{dat, datname, parameter}}

\item{order}{Sort output$gof by RMSE? If FALSE, the order of appearance in 
selection (or dlf$parameter) is kept. DEFAULT: TRUE}

\item{plot}{Call \code{\link{plotLgof}}? DEFAULT: TRUE}

\item{progbars}{Show progress bars for each loop? DEFAULT: TRUE if n > 200}

\item{ks}{Include ks.test results in \code{dlf$gof}?
Computing is much faster when FALSE. DEFAULT: FALSE}

\item{weightc}{Custom weights, see \code{\link{distLweights}}. DEFAULT: NA}

\item{quiet}{Suppress notes? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{plotLgof}}}
}
\value{
List as explained in \code{\link{extremeStat}}. The added element is gof,\cr
a data.frame with the root mean square error (RMSE) and R squared (R2),\cr
if ks=TRUE, the p and D values from a simple ks.test,\cr
as well as weights for each distribution function as explained in \code{\link{distLweights}}
}
\description{
Calculate goodness of fit for several distributions, plot rank comparison.
}
\note{
If you get a \code{note in distLgof: NAs removed in CDF ...}, this
      probably means that the support of some of the fitted distributions do not
      span the whole data range. Instead the outside support regions get NAs that
      are then detected by rmse and rsquare. I plan to fix this with WHA's new supdist.
}
\examples{

library(lmomco)
data(annMax)

# Goodness of Fit is measured by RMSE of cumulated distribution function and ?ecdf
dlf <- distLfit(annMax, cdf=TRUE, nbest=17)
plotLfit(dlf, cdf=TRUE, sel=c("wak", "revgum"))
dlf$gof
x <- sort(annMax)
segments(x0=x, y0=plmomco(x, dlf$parameter$revgum), y1=ecdf(annMax)(x), col=2)
segments(x0=x, y0=plmomco(x, dlf$parameter$wak), y1=ecdf(annMax)(x), col=4, lwd=2)
plot(ecdf(annMax), add=TRUE)
# RMSE: root of average of ( errors squared )  ,   errors = line distances

# weights by three different weighting schemes
plotLgof(dlf)
distLgof(dlf, ks=TRUE, plot=FALSE)$gof

# Kolmogorov-Smirnov Tests return slightly different values:
ks.test(annMax, "pnorm", mean(annMax), sd(annMax) )$p.value
ks.test(annMax, "cdfnor", parnor(lmoms(annMax)))$p.value

# todo: develop examples below function

# GOF: how well do the distributions fit the original data?
pairs(dlf$gof[,1:3]) # measures of goodness of fit are correlated quite well here.

\dontrun{ ## to save CRAN check computing time

plotLfit(dlf, cdf=TRUE, sel=c("pe3", "rice", "revgum"), order=T)
x <- sort(annMax, decreasing=TRUE)[  1:(0.5*length(annMax))  ]
tcdfs <- plmomco(x,dlf$parameter[["revgum"]])
ecdfs <- ecdf(annMax)(x) # Empirical CDF
plot(x, tcdfs, type="o", col=2)
points(x, ecdfs)
linReg(tcdfs, ecdfs, type="o")
abline(a=0,b=1, lty=3)

} # end dontrun

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2014 + July 2015
}
\seealso{
\code{\link{distLweights}}, \code{\link{plotLgof}}, \code{\link{distLfit}} 
    More complex estimates of quality of fits:
    Fard, M.N.P. and Holmquist, B. (2013, Chilean Journal of Statistics): 
    Powerful goodness-of-fit tests for the extreme value distribution.
    http://chjs.mat.utfsm.cl/volumes/04/01/Fard_Holmquist(2013).pdf
}
\keyword{distribution}
\keyword{hplot}
\keyword{univar}

