% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLprint.R
\name{distLprint}
\alias{distLprint}
\title{print dlf objects}
\usage{
distLprint(dlf, digits = 1)
}
\arguments{
\item{dlf}{List as explained in \code{\link{extremeStat}}}

\item{digits}{number of digits \code{\link{round}}ed to. DEFAULT: 1}
}
\value{
none, prints via \code{\link{message}}.
}
\description{
print list objects created in this package
}
\examples{

# see 
?distLextreme

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2015
}
\seealso{
\code{\link{extremeStat}}
}
\keyword{list}
\keyword{methods}
\keyword{print}

