% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLweights.R
\name{distLweights}
\alias{distLweights}
\title{Compute distribution weights from GOF}
\usage{
distLweights(rmse, order = TRUE, weightc = NA)
}
\arguments{
\item{rmse}{Numeric: (named) vector with goodness of fit values (RMSE)}

\item{order}{Logical: should result be ordered by RMSE? DEFAULT: TRUE}

\item{weightc}{Numeric: Named vector with custom weights. DEFAULT: NA}
}
\value{
data.frame
}
\description{
Determine distribution function weights from RMSE for weighted averages.
}
\examples{

distLweights(c(gum=0.20, wak=0.17, gam=0.21))
distLweights(c(gum=0.20, wak=0.17, gam=0.21), order=FALSE)
distLweights(c(gum=0.20))

# RMSE vs R2 for GOF judgement --------
library(lmomco)
data(annMax)

compranks <- function(d)
{
gofProp <- 0.5
x <- sort(annMax, decreasing=TRUE)[  1:(gofProp*length(annMax))  ]
tcdfs <- plmomco(x,dlf50$parameter[[d]])
ecdfs <- ecdf(annMax)(x) # Empirical CDF
# Root Mean Square Error, R squared:
berryFunctions::linReg(tcdfs, ecdfs, lwd=1, pch=16, main=d, digits=5, xlim=c(0.5, 1),
       ylim=c(0.5, 1), pos1="topleft")
abline(a=0,b=1, lty=3)
c(berryFunctions::rmse(tcdfs, ecdfs), berryFunctions::rsquare(tcdfs, ecdfs))
}

dlf50 <- distLfit(annMax, gofProp=0.5)
dn <- rownames(dlf50$gof)

op <- par(mfrow=c(5,4), mar=rep(1,4), xaxt="n", yaxt="n")
for(i in dn) compranks(i)
par(op)
# so revgum, nor and rice systematically deviate from ECDF.
# RMSE is better to sort by than R2

# custom weights
cw <- c("gpa"=7, "gev"=3, "wak"=6, "wei"=4, "kap"=3.5, "gum"=3, "ray"=2.1,
        "ln3"=2, "pe3"=2.5, "gno"=4, "gam"=5)
dlf <- distLfit(annMax, plot=FALSE, weightc=cw)
distLgofPlot(dlf, ranks=TRUE)


}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2016
}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\keyword{distribution}

