% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLweights.R
\name{distLweights}
\alias{distLweights}
\title{Compute distribution weights from GOF}
\usage{
distLweights(rmse, order = TRUE, weightc = NA)
}
\arguments{
\item{rmse}{Numeric: Named vector with goodness of fit values (RMSE).
Can also be a data.frame, in which case the column rmse or RMSE is used.}

\item{order}{Logical: should result be ordered by RMSE? DEFAULT: TRUE}

\item{weightc}{Optional: a named vector with custom weights for each distribution.
Are internally normalized to sum=1 after removing nonfitted dists.
Names match the parameter names from \code{rmse}.
DEFAULT: NA}
}
\value{
data.frame
}
\description{
Determine distribution function weights from RMSE for weighted averages.
The weights are inverse to RMSE: weight1 for all dists, 
weight2 places zero weight on the worst fitting function, 
weight3 on the worst half of functions.
}
\examples{
# weights from RMSE vector:
rmse <- c(gum=0.20, wak=0.17, gam=0.21, gev=0.15)
distLweights(rmse)
distLweights(rmse, order=FALSE)

# weights from RMSE in data.frame:
df <- data.frame("99.9\%"=2:5, rmse=sample(3:6))
rownames(df) <- letters[1:4]
df ;  distLweights(df)

# custom weights:
set.seed(42); x <- data.frame(A=1:5, RMSE=runif(5)) ; x
distLweights(x)
distLweights(x, weightc=c("1"=3, "3"=5)) 
distLweights(x, weightc=c("1"=3, "3"=5), order=FALSE) 

# real life example:
data(annMax)
cw <- c("gpa"=7, "gev"=3, "wak"=6, "wei"=4, "kap"=3.5, "gum"=3, "ray"=2.1,
        "ln3"=2, "pe3"=2.5, "gno"=4, "gam"=5)
dlf <- distLfit(annMax, plot=FALSE, weightc=cw, quiet=TRUE, order=FALSE)
plotLgof(dlf)

# GOF judgement by RMSE, not R2 --------
# Both RMSE and R2 are computed with ECDF and TCDF
# R2 may be very good (see below), but fit needs to be close to 1:1 line, 
# which is better measured by RMSE
set.seed(42); x <- runif(30)
y1 <-     x+rnorm(30,sd=0.09)
y2 <- 1.5*x+rnorm(30,sd=0.01)-0.3
plot(x,x, asp=1, las=1)
berryFunctions::linReg(x, y2, add=TRUE, digits=4, pos1="topleft")
points(x,y2, col="red", pch=3)
points(x,y1, col="blue")
berryFunctions::linReg(x, y1, add=TRUE, digits=4, col="blue", pos1="left")
abline(a=0, b=1, lwd=3, lty=3)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2016
}
\seealso{
\code{\link{distLgof}}, \code{\link{distLquantile}}
}
\keyword{distribution}

