% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{extremeStat-deprecated}
\alias{distLextremePlot}
\alias{distLgof}
\alias{distLgofPlot}
\alias{distLplot}
\alias{distLprint}
\alias{extremeStat-deprecated}
\title{Deprecated Functions}
\usage{
distLextremePlot(...)

distLgof(...)

distLgofPlot(...)

distLplot(...)

distLprint(...)
}
\arguments{
\item{\dots}{Ignored arguments}
}
\description{
With version 1.3.0 (jan 2017), a number of functions have been renamed and restructured.
The computing functions don't plot anymore.
See the new structure in \code{\link{extremeStat}}.
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2017
}
\seealso{
\code{\link{extremeStat}}
}

